requirements.txt
python-telegram-bot
requests
beautifulsoup4
apscheduler
psycopg2
SQLAlchemy
pyTelegramBotAPI